refactor(test): extract AbstractMysqlSyncIT base class and add unit tests
- Extract AbstractMysqlSyncIT from SchemaEvolutionIT/WholeDatabaseSyncIT
  (~200 lines of shared MySQL container, Flink pipeline, and JDBC helpers)
- Add CDBBatchSink tests: Record construction, invoke filtering
- Add SchemaEvolver test: concurrent event processing (thread safety)
- Total unit tests: 67 → 72
